Move matcher and tests from simplify_algebra to simplify_reshapes
Do this to reuse some of the reshape matcher sthat already handle most of the functionality that this matcher uses.
