+++ src/fuse_horizontal.cpp
@@ -275,10 +275,11 @@ struct gather_horizontal_fusion
 //
 // Downstream pointwise ops (SiLU, GeLU, LayerNorm, etc.) are NOT handled
 // here — they are generically lifted above the slice/squeeze boundaries by
-// find_squeeze_splits in simplify_algebra, which runs after this pass in
-// the pipeline.  This separation keeps fuse_horizontal pattern-agnostic
-// while simplify_algebra's iterative framework handles arbitrary activation
-// patterns via repeated pointwise lifting.
+// find_slice_squeeze in simplify_reshapes (which moves the squeeze past
+// the op) together with find_splits in simplify_algebra (which merges the
+// parallel branches).  This separation keeps fuse_horizontal
+// pattern-agnostic while the iterative framework handles arbitrary
+// activation patterns.
 // ---------------------------------------------------------------------------
 
 struct dot_horizontal_fusion
