+++ test/simplify_algebra_test.cpp
@@ -5074,287 +5074,4 @@ TEST_CASE(pow3)
     EXPECT(m1.sort() == m2.sort());
 }
 
-// ---------------------------------------------------------------------------
-// find_squeeze_splits: lifts pointwise ops above axis-0 slice→squeeze
-// compound splits produced by dot_horizontal_fusion.
-// ---------------------------------------------------------------------------
-
-TEST_CASE(squeeze_split_unary_lift)
-{
-    migraphx::shape s{migraphx::shape::float_type, {2, 4}};
-    migraphx::module m1;
-    {
-        auto input = m1.add_parameter("input", s);
-        auto s0    = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), input);
-        auto sq0  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto rel0 = m1.add_instruction(migraphx::make_op("relu"), sq0);
-        auto s1   = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), input);
-        auto sq1  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-        auto rel1 = m1.add_instruction(migraphx::make_op("relu"), sq1);
-        m1.add_return({rel0, rel1});
-    }
-    run_pass(m1);
-
-    migraphx::module m2;
-    {
-        auto input = m2.add_parameter("input", s);
-        auto relu  = m2.add_instruction(migraphx::make_op("relu"), input);
-        auto s0    = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), relu);
-        auto sq0 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto s1  = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), relu);
-        auto sq1 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-        m2.add_return({sq0, sq1});
-    }
-    EXPECT(m1.sort() == m2.sort());
-}
-
-TEST_CASE(squeeze_split_unary_lift_3d)
-{
-    migraphx::shape s{migraphx::shape::float_type, {3, 2, 4}};
-    migraphx::module m1;
-    {
-        auto input = m1.add_parameter("input", s);
-        auto s0    = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), input);
-        auto sq0  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto rel0 = m1.add_instruction(migraphx::make_op("relu"), sq0);
-        auto s1   = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), input);
-        auto sq1  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-        auto rel1 = m1.add_instruction(migraphx::make_op("relu"), sq1);
-        auto s2   = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {2}}, {"ends", {3}}}), input);
-        auto sq2  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s2);
-        auto rel2 = m1.add_instruction(migraphx::make_op("relu"), sq2);
-        m1.add_return({rel0, rel1, rel2});
-    }
-    run_pass(m1);
-
-    migraphx::module m2;
-    {
-        auto input = m2.add_parameter("input", s);
-        auto relu  = m2.add_instruction(migraphx::make_op("relu"), input);
-        auto s0    = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), relu);
-        auto sq0 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto s1  = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), relu);
-        auto sq1 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-        auto s2  = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {2}}, {"ends", {3}}}), relu);
-        auto sq2 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s2);
-        m2.add_return({sq0, sq1, sq2});
-    }
-    EXPECT(m1.sort() == m2.sort());
-}
-
-TEST_CASE(squeeze_split_binary_const_lift)
-{
-    migraphx::shape s{migraphx::shape::float_type, {2, 4}};
-    migraphx::shape bs{migraphx::shape::float_type, {4}};
-    migraphx::module m1;
-    {
-        auto input = m1.add_parameter("input", s);
-        auto b0    = m1.add_literal(migraphx::generate_literal(bs, 0));
-        auto b1    = m1.add_literal(migraphx::generate_literal(bs, 1));
-
-        auto s0 = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), input);
-        auto sq0  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto add0 = m1.add_instruction(migraphx::make_op("add"), sq0, b0);
-
-        auto s1 = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), input);
-        auto sq1  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-        auto add1 = m1.add_instruction(migraphx::make_op("add"), sq1, b1);
-
-        m1.add_return({add0, add1});
-    }
-    run_pass(m1);
-
-    migraphx::module m2;
-    {
-        auto input   = m2.add_parameter("input", s);
-        auto b0      = m2.add_literal(migraphx::generate_literal(bs, 0));
-        auto b1      = m2.add_literal(migraphx::generate_literal(bs, 1));
-        auto b0_unsq = m2.add_instruction(migraphx::make_op("unsqueeze", {{"axes", {0}}}), b0);
-        auto b1_unsq = m2.add_instruction(migraphx::make_op("unsqueeze", {{"axes", {0}}}), b1);
-        auto stacked =
-            m2.add_instruction(migraphx::make_op("concat", {{"axis", 0}}), b0_unsq, b1_unsq);
-        auto add = m2.add_instruction(migraphx::make_op("add"), input, stacked);
-
-        auto s0 = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), add);
-        auto sq0 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto s1  = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), add);
-        auto sq1 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-
-        m2.add_return({sq0, sq1});
-    }
-    EXPECT(m1.sort() == m2.sort());
-}
-
-TEST_CASE(squeeze_split_binary_two_roots)
-{
-    migraphx::shape s{migraphx::shape::float_type, {2, 4}};
-    migraphx::module m1;
-    {
-        auto a = m1.add_parameter("a", s);
-        auto b = m1.add_parameter("b", s);
-
-        auto sa0 = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), a);
-        auto sqa0 = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), sa0);
-        auto sb0  = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), b);
-        auto sqb0 = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), sb0);
-        auto mul0 = m1.add_instruction(migraphx::make_op("mul"), sqa0, sqb0);
-
-        auto sa1 = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), a);
-        auto sqa1 = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), sa1);
-        auto sb1  = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), b);
-        auto sqb1 = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), sb1);
-        auto mul1 = m1.add_instruction(migraphx::make_op("mul"), sqa1, sqb1);
-
-        m1.add_return({mul0, mul1});
-    }
-    run_pass(m1);
-
-    migraphx::module m2;
-    {
-        auto a   = m2.add_parameter("a", s);
-        auto b   = m2.add_parameter("b", s);
-        auto mul = m2.add_instruction(migraphx::make_op("mul"), a, b);
-
-        auto s0 = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), mul);
-        auto sq0 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto s1  = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), mul);
-        auto sq1 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-
-        m2.add_return({sq0, sq1});
-    }
-    EXPECT(m1.sort() == m2.sort());
-}
-
-TEST_CASE(squeeze_split_silu_chain)
-{
-    migraphx::shape s{migraphx::shape::float_type, {2, 1, 4}};
-    migraphx::shape bs{migraphx::shape::float_type, {1, 4}};
-    migraphx::module m1;
-    {
-        auto input = m1.add_parameter("input", s);
-        auto b0    = m1.add_literal(migraphx::generate_literal(bs, 0));
-        auto b1    = m1.add_literal(migraphx::generate_literal(bs, 1));
-
-        auto s0 = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), input);
-        auto sq0  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto add0 = m1.add_instruction(migraphx::make_op("add"), sq0, b0);
-        auto sig0 = m1.add_instruction(migraphx::make_op("sigmoid"), add0);
-        auto mul0 = m1.add_instruction(migraphx::make_op("mul"), sig0, add0);
-
-        auto s1 = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), input);
-        auto sq1  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-        auto add1 = m1.add_instruction(migraphx::make_op("add"), sq1, b1);
-        auto sig1 = m1.add_instruction(migraphx::make_op("sigmoid"), add1);
-        auto mul1 = m1.add_instruction(migraphx::make_op("mul"), sig1, add1);
-
-        m1.add_return({mul0, mul1});
-    }
-    run_pass(m1);
-
-    migraphx::module m2;
-    {
-        auto input   = m2.add_parameter("input", s);
-        auto b0      = m2.add_literal(migraphx::generate_literal(bs, 0));
-        auto b1      = m2.add_literal(migraphx::generate_literal(bs, 1));
-        auto b0_unsq = m2.add_instruction(migraphx::make_op("unsqueeze", {{"axes", {0}}}), b0);
-        auto b1_unsq = m2.add_instruction(migraphx::make_op("unsqueeze", {{"axes", {0}}}), b1);
-        auto stacked =
-            m2.add_instruction(migraphx::make_op("concat", {{"axis", 0}}), b0_unsq, b1_unsq);
-        auto add = m2.add_instruction(migraphx::make_op("add"), input, stacked);
-        auto sig = m2.add_instruction(migraphx::make_op("sigmoid"), add);
-        auto mul = m2.add_instruction(migraphx::make_op("mul"), sig, add);
-
-        auto s0 = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), mul);
-        auto sq0 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto s1  = m2.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), mul);
-        auto sq1 = m2.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-
-        m2.add_return({sq0, sq1});
-    }
-    EXPECT(m1.sort() == m2.sort());
-}
-
-TEST_CASE(squeeze_split_no_fusion_single_split)
-{
-    migraphx::shape s{migraphx::shape::float_type, {1, 4}};
-    migraphx::module m1;
-    {
-        auto input = m1.add_parameter("input", s);
-        auto s0    = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), input);
-        auto sq0  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto rel0 = m1.add_instruction(migraphx::make_op("relu"), sq0);
-        m1.add_return({rel0});
-    }
-    migraphx::module m2 = m1;
-    run_pass(m1);
-    EXPECT(m1.sort() == m2.sort());
-}
-
-TEST_CASE(squeeze_split_no_fusion_gap)
-{
-    migraphx::shape s{migraphx::shape::float_type, {3, 4}};
-    migraphx::module m1;
-    {
-        auto input = m1.add_parameter("input", s);
-        auto s0    = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), input);
-        auto sq0  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto rel0 = m1.add_instruction(migraphx::make_op("relu"), sq0);
-        auto s1   = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {2}}, {"ends", {3}}}), input);
-        auto sq1  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-        auto rel1 = m1.add_instruction(migraphx::make_op("relu"), sq1);
-        m1.add_return({rel0, rel1});
-    }
-    migraphx::module m2 = m1;
-    run_pass(m1);
-    EXPECT(m1.sort() == m2.sort());
-}
-
-TEST_CASE(squeeze_split_no_fusion_different_ops)
-{
-    migraphx::shape s{migraphx::shape::float_type, {2, 4}};
-    migraphx::module m1;
-    {
-        auto input = m1.add_parameter("input", s);
-        auto s0    = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {0}}, {"ends", {1}}}), input);
-        auto sq0  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s0);
-        auto rel0 = m1.add_instruction(migraphx::make_op("relu"), sq0);
-        auto s1   = m1.add_instruction(
-            migraphx::make_op("slice", {{"axes", {0}}, {"starts", {1}}, {"ends", {2}}}), input);
-        auto sq1  = m1.add_instruction(migraphx::make_op("squeeze", {{"axes", {0}}}), s1);
-        auto sig1 = m1.add_instruction(migraphx::make_op("sigmoid"), sq1);
-        m1.add_return({rel0, sig1});
-    }
-    migraphx::module m2 = m1;
-    run_pass(m1);
-    EXPECT(m1.sort() == m2.sort());
-}
-
 int main(int argc, const char* argv[]) { test::run(argc, argv); }
